requirements.txt
torch>=2.2.1
torchvision>=0.17.1
numpy>=1.26.4
PyYAML>=6.0.1
matplotlib>=3.8.2